Move Rewritting functionality to replacement folder
