+++ components/bpmn-q/modeler-component/extensions/quantme/replacement/QuantMETransformator.js
@@ -53,6 +53,9 @@ export async function startQuantmeReplacementProcess(
     };
   }
 
+  // TODO: perform rewriting
+  console.log("Rewriting...")
+
   // get all QuantME modeling constructs from the process
   let replacementConstructs = getQuantMETasks(rootElement, elementRegistry);
   console.log(
